Refactor menuinst functions into subdirectory
